kleinigkeiten

+++ app/controllers/users_controller.rb
@@ -1,5 +1,6 @@
 class UsersController < ApplicationController
 
+  # anmelden eines registrierten benutzers
   def register
     @user = User.find_by(login: params[:login])
 
@@ -10,9 +11,11 @@ class UsersController < ApplicationController
       render json: @user.to_json(only: %w(salt_masterkey privkey_user_enc pubkey_user))
     end
   end
-
+# anlegegen eines neuen benutzers
   def create
-    @user = User.new(login: params[:login], salt_masterkey: params[:salt_masterkey], pubkey_user: params[:pubkey_user],
+    @user = User.new(login: params[:login],
+                     salt_masterkey: params[:salt_masterkey],
+                     pubkey_user: params[:pubkey_user],
                      privkey_user_enc: params[:privkey_user_enc])
 
     if @user.save
@@ -21,15 +24,17 @@ class UsersController < ApplicationController
       head 400
     end
   end
-
+# user loeschen
   def destroy
     user = User.check_sig(params[:timestamp].to_s, params[:login], params[:digitale_signatur])
-
-    user.destroy
-
-    head 200
+    if user.nil?
+      head 400
+    else
+      user.destroy
+      head 200
+    end
   end
-
+# anfordnern des pubkeys einer benutzers
   def pubkey
     @user = User.find_by(login: params[:login])
